api/config.py
import boto3
from botocore.exceptions import ClientError
import json

class SecretsManager:
    def __init__(self, region_name="us-east-1"):
        self.session = boto3.session.Session()
        self.client = self.session.client(
            service_name='secretsmanager',
            region_name=region_name
        )

    def get_secret(self, secret_name):
        try:
            get_secret_value_response = self.client.get_secret_value(
                SecretId=secret_name
            )
        except ClientError as e:
            raise e
        else:
            if 'SecretString' in get_secret_value_response:
                secret = json.loads(get_secret_value_response['SecretString'])
                return secret
            else:
                raise ValueError("Secret value is not a string")